vae/minimax_h3: drop what the measurements did not justify
Review pass over the VAE half of this PR. Net -131 lines.

Removed the tracing machinery entirely -- `trace_device_stitch`, `_stitch_graph`, the `Tracer`
wiring and the harness's `traced` mode. It measured no benefit (6.887 s traced against 6.934 s
untraced at 768P/15s), its output correctness was never verified, and nothing exercised it since
it defaulted off. Its `trace_region_size` was also being imposed on every other harness mode,
quietly changing what they measured. The finding it produced -- that the stage is device-bound and
host dispatch was already hidden underneath -- is recorded in `MiniMaxH3.md` instead, where it
cannot rot against the code.

Removed `MINIMAX_H3_TIME_DISPATCH` for the same reason: it is instrumentation whose second number
(the post-synchronize tail, not total device time) invites exactly the wrong conclusion. The
pitfall is documented in `MiniMaxH3.md`.

Collapsed `_enqueue_clip_device_stitched` / `_read_stitched_canvas` back into
`_decode_clip_device_stitched`. The split existed only to support a readback/compute overlap that
measured as a pessimisation and was reverted; each half had exactly one caller.

Dropped `DeviceTileStitcher`'s `layout` parameter. Neither construction site passed it, so it was
configurability nobody used; the ROW_MAJOR it defaulted to is now simply what the class does, with
the reasoning moved to the class docstring where it applies to the whole blend rather than to one
dead argument.

Moved the whole-stage decode timing out of `tools/decode_perf_harness.py` and into
`test_performance_vae_minimax_h3.py`, next to `test_visual_data_parallel_throughput` -- the way LTX
keeps its decode tests in `test_vae_ltx.py` rather than in a separate harness file.

`common.py` now names its mesh shapes and composes them into a list, following
`ltx_mesh_params.py`, instead of building params from a dict through a `mesh_params()` helper whose
only call site passed no arguments.

Also drops the commentary around the `local_device_to_torch` calls; the call is the point.
