+++ models/tt_dit/tests/models/minimax_h3/test_performance_vae_minimax_h3.py
@@ -16,6 +16,8 @@ import ttnn
 
 from ....models.audio_vae.minimax_h3.convert_minimax_h3_audio import convert_minimax_h3_audio_state_dict
 from ....models.audio_vae.minimax_h3.encoder_minimax_h3_audio import MiniMaxH3AudioEncoder
+from ....models.vae.minimax_h3.vae_minimax_h3 import MiniMaxH3Vae, MiniMaxH3VaeConfig
+from ....parallel.manager import CCLManager
 from .common import (
     CLIP_FRAMES,
     DECODE_LATENT_FRAMES,
@@ -195,3 +197,131 @@ def test_audio_baselines(mesh_device):
         "audio_decode_5s": _time_it(lambda: decoder(latents), iterations=1, mesh_device=mesh_device),
     }
     _report(measurements)
+
+
+# ---- whole-stage `decode` timing, per output type -------------------------------------------
+#
+# `test_visual_data_parallel_throughput` times one decoder wave; this times the stage the pipeline
+# bills to "VAE decode" -- tiling, upload, readback, unpatchify, stitch and the temporal assembly --
+# and prints `_report_profile`'s breakdown, which is the point: a stage total alone cannot say which
+# of device / readback / stitch moved.
+#
+# RING, not the `fabric_config=None` the throughput test uses: `_decode_clip_device_stitched`
+# hardcodes `all_gather(topology=Ring)`, so `yuv420` cannot run on a line fabric. Built with a
+# `CCLManager` because `_read_wave_units` picks its reader on `ccl_manager is None` -- without one a
+# single-host run silently takes the `ConcatMeshToTensor` path and measures something production
+# never runs.
+#
+#   export MINIMAX_H3_MODEL_PATH=/path/to/MiniMax-H3
+#   pytest models/tt_dit/tests/models/minimax_h3/test_performance_vae_minimax_h3.py -k decode_stage -s
+
+MESH_4X8_RING = [
+    pytest.param(
+        (4, 8),
+        {
+            "fabric_config": ttnn.FabricConfig.FABRIC_1D_RING,
+            "require_exact_physical_num_devices": True,
+            "l1_small_size": 65536,
+        },
+        id="4x8ring",
+    )
+]
+
+# 1344x768 at 16x spatial compression. Latent frame counts follow the 17n -> 5n rule less
+# `token_drop`, so 124 frames (5 s) is 37 and 362 (15 s) is 107.
+DECODE_STAGE_LATENT_HW = (48, 84)
+DECODE_STAGE_FRAMES = {5: 37, 15: 107}
+# 7 latent frames is one chunk, the cheapest warm-up that still exercises a whole wave.
+DECODE_STAGE_WARM_FRAMES = 7
+
+MINIMAX_H3_PIXEL_MEAN = (0.485, 0.456, 0.406)
+MINIMAX_H3_PIXEL_STD = (0.229, 0.224, 0.225)
+
+# Each mode is one lever, so a row of this table measures that lever alone.
+DECODE_STAGE_MODES = {
+    # Today's default: bf16 tiles in token space, blended and unpatchified on host.
+    "float": {},
+    # Same path, uint8 across PCIe. Halves the transfer.
+    "uint8": {"pixel_denorm": (MINIMAX_H3_PIXEL_MEAN, MINIMAX_H3_PIXEL_STD), "readback_uint8": True},
+    # Stitch, clamp and colour-convert on device; read one planar canvas at 1.5 bytes/pixel.
+    "yuv420": {"pixel_denorm": (MINIMAX_H3_PIXEL_MEAN, MINIMAX_H3_PIXEL_STD), "device_stitch": True},
+}
+
+
+def _decode_stage_state(weights_dir: str) -> dict[str, torch.Tensor]:
+    """Just the decoder-side tensors; over half the checkpoint is the encoder."""
+    import json
+
+    from safetensors.torch import load_file
+
+    index_path = os.path.join(weights_dir, "diffusion_pytorch_model.safetensors.index.json")
+    if not os.path.isfile(index_path):
+        return {
+            k: v
+            for k, v in load_file(os.path.join(weights_dir, "diffusion_pytorch_model.safetensors")).items()
+            if k.startswith(("decoder.", "post_quant_conv."))
+        }
+    weight_map = json.loads(open(index_path).read())["weight_map"]
+    wanted = {k: f for k, f in weight_map.items() if k.startswith(("decoder.", "post_quant_conv."))}
+    state: dict[str, torch.Tensor] = {}
+    for shard in sorted(set(wanted.values())):
+        loaded = load_file(os.path.join(weights_dir, shard))
+        state.update({k: loaded[k] for k in wanted if k in loaded})
+    return state
+
+
+@pytest.mark.timeout(3600)
+@pytest.mark.parametrize(("mesh_device", "device_params"), MESH_4X8_RING, indirect=["mesh_device", "device_params"])
+@pytest.mark.parametrize("seconds", [5, 15], ids=["5s", "15s"])
+@pytest.mark.parametrize("mode", sorted(DECODE_STAGE_MODES), ids=sorted(DECODE_STAGE_MODES))
+def test_decode_stage(mesh_device, seconds, mode):
+    """Measurement driver, not a gate: reports the split, asserts nothing about wall time."""
+    from loguru import logger
+
+    weights_dir = weights_subdir("vae")
+    if weights_dir is None:
+        pytest.skip("MiniMax-H3 vae not found; set MINIMAX_H3_MODEL_PATH")
+
+    config = MiniMaxH3VaeConfig.from_pretrained(weights_dir)
+    ccl_manager = CCLManager(mesh_device, num_links=2, topology=ttnn.Topology.Ring)
+    vae = MiniMaxH3Vae(config, mesh_device=mesh_device, ccl_manager=ccl_manager, **DECODE_STAGE_MODES[mode])
+    output_type = "yuv420" if mode == "yuv420" else "float"
+
+    vae.load_decoder_state(_decode_stage_state(weights_dir))
+    # Build the per-shape decoder outside the timed region, as `_prepare_vae(decode_shape=...)` does:
+    # its weight upload is seconds and would otherwise land inside the first `decode`.
+    vae._decoder_for(config.tokens_chunk_size + config.token_overlap, 16, 16)
+
+    torch.manual_seed(0)
+    latent_h, latent_w = DECODE_STAGE_LATENT_HW
+    warm = torch.randn(1, config.latent_channels, DECODE_STAGE_WARM_FRAMES, latent_h, latent_w)
+    latents = torch.randn(1, config.latent_channels, DECODE_STAGE_FRAMES[seconds], latent_h, latent_w)
+
+    vae.decode(warm, output_type=output_type)
+
+    started = time.time()
+    video = vae.decode(latents, output_type=output_type)
+    elapsed = time.time() - started
+
+    profile = dict(vae.last_decode_profile)
+    logger.info(f"DECODE_STAGE mode={mode} seconds={seconds} total={elapsed:.3f}s video={tuple(video.shape)}")
+    logger.info(
+        f"DECODE_SPLIT mode={mode} seconds={seconds} "
+        + " ".join(
+            f"{k}={profile.get(k, 0.0):.3f}"
+            for k in (
+                "device",
+                "readback",
+                "stitch",
+                "unpatchify",
+                "blend",
+                "concat",
+                "tiling",
+                "upload",
+                "host_prep",
+                "residual",
+            )
+        )
+        + f" waves={int(profile.get('waves', 0))} units={int(profile.get('units', 0))}"
+        + f" readback_gb={profile.get('readback_mb', 0.0) / 1000:.2f}"
+    )
